test:전략 패턴을 위해 곱셈 테스트를 StringCalculatorTest에서 MultiplyOperatorTest로 리팩토링
